+++ documentation/src/components/landing/section-no-constraints/index.tsx
@@ -1194,8 +1194,8 @@ export const SectionNoConstraints: React.FC<{ starCount?: number }> = ({
                                                 It has a very strong community
                                                 of maintainers, contributers and
                                                 and users providing{" "}
-                                                <strong>7/24</strong> support on
-                                                our Github, Twitter and Discord
+                                                <strong>24/7</strong> support on
+                                                our GitHub, Twitter and Discord
                                                 channels.
                                             </p>
                                         </div>