+++ packages/antd/src/components/pages/auth/components/index.tsx
@@ -1,4 +1,4 @@
 export { LoginPage } from "./login";
 export { RegisterPage } from "./register";
-export { ResetPasswordPage } from "./resetPassword";
+export { ForgotPasswordPage } from "./forgotPassword";
 export { UpdatePasswordPage } from "./updatePassword";