+++ packages/antd/src/components/pages/auth/components/register/index.tsx
@@ -15,6 +15,7 @@ import {
     LayoutProps,
     CardProps,
     FormProps,
+    Divider,
 } from "antd";
 import {
     useTranslate,
@@ -34,11 +35,11 @@ type RegisterProps = RefineRegisterPageProps<LayoutProps, CardProps, FormProps>;
  * @see {@link https://refine.dev/docs/ui-frameworks/antd/components/antd-auth-page/#register} for more details.
  */
 export const RegisterPage: React.FC<RegisterProps> = ({
+    providers,
     loginLink,
     wrapperProps,
     contentProps,
     renderContent,
-    onSubmit,
     formProps,
 }) => {
     const [form] = Form.useForm<RefineRegisterFormTypes>();
@@ -54,6 +55,41 @@ export const RegisterPage: React.FC<RegisterProps> = ({
         </Title>
     );
 
+    const renderProviders = () => {
+        if (providers) {
+            return (
+                <>
+                    {providers.map((provider) => {
+                        return (
+                            <Button
+                                key={provider.name}
+                                type="ghost"
+                                block
+                                icon={provider.icon}
+                                style={{
+                                    display: "flex",
+                                    justifyContent: "center",
+                                    alignItems: "center",
+                                    width: "100%",
+                                    marginBottom: "8px",
+                                }}
+                                onClick={() =>
+                                    register({
+                                        providerName: provider.name,
+                                    })
+                                }
+                            >
+                                {provider.label}
+                            </Button>
+                        );
+                    })}
+                    <Divider>{translate("pages.login.divider", "or")}</Divider>
+                </>
+            );
+        }
+        return null;
+    };
+
     const CardContent = (
         <Card
             title={CardTitle}
@@ -61,10 +97,11 @@ export const RegisterPage: React.FC<RegisterProps> = ({
             style={containerStyles}
             {...(contentProps ?? {})}
         >
+            {renderProviders()}
             <Form<RefineRegisterFormTypes>
                 layout="vertical"
                 form={form}
-                onFinish={(values) => (onSubmit ?? register)(values)}
+                onFinish={(values) => register(values)}
                 requiredMark={false}
                 {...formProps}
             >
@@ -135,15 +172,17 @@ export const RegisterPage: React.FC<RegisterProps> = ({
                     )}
                 </div>
 
-                <Button
-                    type="primary"
-                    size="large"
-                    htmlType="submit"
-                    loading={isLoading}
-                    block
-                >
-                    {translate("pages.register.buttons.submit", "Sign up")}
-                </Button>
+                <Form.Item>
+                    <Button
+                        type="primary"
+                        size="large"
+                        htmlType="submit"
+                        loading={isLoading}
+                        block
+                    >
+                        {translate("pages.register.buttons.submit", "Sign up")}
+                    </Button>
+                </Form.Item>
             </Form>
         </Card>
     );