+++ packages/mantine/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pankod/refine-mantine
 
+## 1.0.6
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @pankod/refine-ui-types@0.11.0
+
+## 1.0.5
+
+### Patch Changes
+
+-   Updated dependencies [[`a65525de6f`](https://github.com/pankod/refine/commit/a65525de6f995babfca1058e933cdbea67d6032e)]:
+    -   @pankod/refine-ui-types@0.10.0
+
 ## 1.0.4
 
 ### Patch Changes